refactor(server): decompose pr-feedback-service.ts into 3 focused modules
